recieve x or X val as long

+++ ft_printf.c
@@ -113,7 +113,7 @@ char *ft_translate_fmt(t_plist *params, va_list *args)
     else if (params->type == 'u')
         param_str = ft_long_itoa(va_arg(*args, unsigned int));
     else if (params->type == 'x' || params->type == 'X')
-        param_str = ft_convert_to_hex(va_arg(*args, int), params->type == 'X');
+        param_str = ft_convert_to_hex(va_arg(*args, long), params->type == 'X');
     else if (params->type == 'p')
     {
         p = va_arg(*args, void*);
